convert milliseconds to human readable time format removed Link field which is not used Comment by George: First time I tried to make function in same class, however, somehow it shows compile error, so I write code inside of run.complete request statement I'm not sure it is right code structure, but it works for now. I hope I can make this better in future.

+++ src/qase/payload.go
@@ -34,5 +34,5 @@ type CompleteTestPayload struct {
 	Failed   uint8 `json:"failed"`
 	Passed   uint8 `json:"passed"`
 	Blocked  uint8 `json:"blocked"`
-	Duration int64 `json:"duration"`
+	Duration int   `json:"duration"`
 }
